requirements.txt
python-dotenv~=1.0.1
requests~=2.31.0
DateTime~=5.4
beautifulsoup4~=4.12.3